% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safran2raster.R
\name{DT2rast}
\alias{DT2rast}
\title{Transforme grille Safran en raster simple}
\usage{
DT2rast(DT, varname, extr_date)
}
\arguments{
\item{DT}{data.table object with Safran data}

\item{varname}{Single character with variable name}

\item{extr_date}{Single character with date (format : "YYYYMMDD")}
}
\value{
Raster layer
}
\description{
Filtre data.table avec 1 variable et 1 date
et génère 1 raster (toute l'étendue c'est à dire la Métropole)
}
\examples{
f_safran <- system.file("extdata", "dumdata_287009.csv.gz", package = "safranpack")
DT_safran <- fread(f_safran, header = TRUE)
raster_rr <- DT2rast(DT_safran, "PRELIQ_Q", "28700918")
writeRaster(raster_rr, "rr_france_28700918.tif")

}
