% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safran2raster.R
\name{safran2rast}
\alias{safran2rast}
\title{Transforme grille Safran en raster multidates}
\usage{
safran2rast(DT, varname, d_date, f_date, etendue_l2e = NULL)
}
\arguments{
\item{DT}{data.table object with Safran data}

\item{varname}{Single character with variable name}

\item{d_date}{Single character with begin date (format : "YYYYMMDD")}

\item{f_date}{Single character with end date (format : "YYYYMMDD")}

\item{etendue_l2e}{Extent Lambert2E}
}
\value{
Raster stack
}
\description{
Filtre data.table avec 1 variable, 1 plage temporelle et 1 étendue spatiale
et genere 1 stack raster
}
\examples{
f_safran <- system.file("extdata", "dumdata_287009.csv.gz", package = "safranpack")
#DT_safran <- fread(f_safran, header = TRUE)
#raster_rr_287009 <- safran2rast(DT_safran, "PRELIQ_Q", "28700901", "28700930")
#writeRaster(raster_rr_287009, "rr_france_287009.tif")

}
