% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrer_safran.R
\name{safran_filtre}
\alias{safran_filtre}
\title{Filtrer données Safran par lieu (XY) et plage temporelle}
\usage{
safran_filtre(DT, lamb_x, lamb_y, date_d, date_f)
}
\arguments{
\item{DT}{data.table object with Safran data}

\item{lamb_x}{X lambert 2 etendu}

\item{lamb_y}{Y lambert 2 etendu}

\item{date_d}{Single character with begin date (format : "YYYYMMDD")}

\item{date_f}{Single character with end date (format : "YYYYMMDD")}
}
\value{
data.table object with Safran data
}
\description{
Filtrer données Safran par lieu (XY) et plage temporelle
}
\examples{
DT_XY <- safran_filtre(DT, "640", "2200", "20180301", "20180331")

}
